chore(module:carousel): refactor
