+++ components/carousel/carousel.module.ts
@@ -9,8 +9,8 @@ import { PlatformModule } from '@angular/cdk/platform';
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 
-import { NzCarouselContentDirective } from './nz-carousel-content.directive';
-import { NzCarouselComponent } from './nz-carousel.component';
+import { NzCarouselContentDirective } from './carousel-content.directive';
+import { NzCarouselComponent } from './carousel.component';
 
 @NgModule({
   declarations: [NzCarouselComponent, NzCarouselContentDirective],
